Fix iterating through previous rcon commands #5
When iterating through old commands you entered previously in the rcon command text input using the arrow up/down keys, you'd have to press the new direction twice to get the next item when changing directions.
This is annoying, so keep track of if we're switching directions and skip the value that's currently displayed in the underlying iterator.

Fixes #5